\name{erpR-package}
\alias{erpR-package}
\alias{erpR}
\docType{package}
\title{
erpR-package
}
\description{
The erpR package includes a series of functions for computing traditional ERP measure, plotting, importing ERP data and for exploratory analyses.
}
\details{
\tabular{ll}{
Package: \tab erpR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-19\cr
License: \tab GNU public license \cr
}
The package includes a series of functions for computing traditional ERP measure, plotting, importing ERP data and for exploratory analyses.
\author{
Giorgio Acara

\email<giorgio.arcara@gmail.com>

Mantainer: giorgio.arcara@gmail.com
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
library(erpR)
}
